requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
langchain-core
langchain-community
langchain-groq
sqlmodel